Fixed shooting and some small stuff (first Android build)

+++ Assets/Scripts/Character.cs
@@ -15,7 +15,8 @@ public class Character : ScriptableObject {
     public int health;
     public int speed;
     public int damage;
-    public int weaponRange;
+
+    public float weaponRange;
 
     public bool isWeaponRay;
